# Inventory-Service/src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: BookYourShow - Inventory API
  description: REST APIs for managing inventory (movies, theatres, shows, seats) in BookYourShow Application
  version: 1.0.0
  contact:
    name: bookYourShow Team
    email: bookyourshow@company.com
...
servers:
  - url: http://localhost:8080/inventory/v1
    description: dev
paths:
...
  /theaters:
    get:
      tags: [Theatres]
      summary: Get theatres by city
      operationId: getTheatres
      parameters:
        - in: query
          name: cityId
          required: true
          description: City ID
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: List of theaters in the city
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TheatreListResponse'
        '400':
          description: Invalid city
        '500':
          description: Internal server error
    post:
      tags: [Theatres]
      summary: Create a new theatre (admin only)
      operationId: createTheatre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TheatreCreateRequest'
      responses:
        '201':
          description: Theatre created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theatre'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: City or owner not found
  /theaters/{theatreId}:
    put:
      tags: [Theatres]
      summary: Update a theatre (admin or theatre owner)
      operationId: updateTheatre
      parameters:
        - in: path
          name: theatreId
          required: true
          description: Unique theatre ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-b000-000000000001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TheatreUpdateRequest'
      responses:
        '200':
          description: Theatre updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theatre'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Theatre not found
    delete:
      tags: [Theatres]
      summary: Delete a theatre (admin only)
      operationId: deleteTheatre
      parameters:
        - in: path
          name: theatreId
          required: true
          description: Unique theatre ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-b000-000000000001"
      responses:
        '204':
          description: Theatre deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Theatre not found
        '409':
          description: Theatre still has screens or shows
  /theaters/{theatreId}/screens:
    post:
      tags: [Theatres]
      summary: Create a new screen for a theatre (admin or theatre owner)
      operationId: createScreen
      parameters:
        - in: path
          name: theatreId
          required: true
          description: Unique theatre ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-b000-000000000001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScreenCreateRequest'
      responses:
        '201':
          description: Screen created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Screen'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Theatre or format not found
  /theaters/{theatreId}/screens/{screenId}:
    put:
      tags: [Theatres]
      summary: Update a screen (admin or theatre owner)
      operationId: updateScreen
      parameters:
        - in: path
          name: theatreId
          required: true
          description: Unique theatre ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-b000-000000000001"
        - in: path
          name: screenId
          required: true
          description: Unique screen ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-c000-000000000001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScreenUpdateRequest'
      responses:
        '200':
          description: Screen updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Screen'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Theatre or screen not found
    delete:
      tags: [Theatres]
      summary: Delete a screen (admin or theatre owner)
      operationId: deleteScreen
      parameters:
        - in: path
          name: theatreId
          required: true
          description: Unique theatre ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-b000-000000000001"
        - in: path
          name: screenId
          required: true
          description: Unique screen ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-c000-000000000001"
      responses:
        '204':
          description: Screen deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Theatre or screen not found
        '409':
          description: Screen still has shows scheduled
  /theaters/{theatreId}/screens/{screenId}/seats:bulk:
    post:
      tags: [Theatres]
      summary: Bulk-create seats for a screen (admin or theatre owner)
      description: >
        Bulk creation since a single screen typically has ~150 seats.
        Provide either `rows` (rowLabel + seatCount per row) for sequential
        seat numbering, or `seats` (explicit rowLabel/seatNumber pairs).
        Rejected if shows already exist for this screen.
      operationId: bulkCreateSeats
      parameters:
        - in: path
          name: theatreId
          required: true
          description: Unique theatre ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-b000-000000000001"
        - in: path
          name: screenId
          required: true
          description: Unique screen ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-c000-000000000001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeatBulkCreateRequest'
      responses:
        '201':
          description: Seats created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatListResponse'
        '400':
          description: Invalid request (must provide either rows or seats; not both empty)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Theatre or screen not found
        '409':
          description: Screen already has shows scheduled or duplicate seats
  /theaters/{theatreId}/screens/{screenId}/seats:
    delete:
      tags: [Theatres]
      summary: Delete all seats for a screen (admin or theatre owner)
      description: Clear seats so they can be re-seeded. Rejected if shows are scheduled.
      operationId: deleteScreenSeats
      parameters:
        - in: path
          name: theatreId
          required: true
          description: Unique theatre ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-b000-000000000001"
        - in: path
          name: screenId
          required: true
          description: Unique screen ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-c000-000000000001"
      responses:
        '204':
          description: Seats deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Theatre or screen not found
        '409':
          description: Screen has shows scheduled
components:
  schemas:
...
    TheatreListResponse:
      type: object
      properties:
        theatres:
          type: array
          items:
            $ref: '#/components/schemas/Theatre'

    # ── City ────────────────────────────────────────────────
...
    Theatre:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-4000-b000-000000000001"
        name:
          type: string
          example: PVR Cinemas
        city:
          type: string
          example: Chennai
        cityId:
          type: integer
          description: City ID this theatre belongs to
          example: 1
        location:
          type: string
          example: Phoenix MarketCity, Velachery
        latitude:
          type: number
          format: double
          example: 12.9716
        longitude:
          type: number
          format: double
          example: 77.5946
        ownerId:
          type: string
          format: uuid
          description: User ID of the theatre owner (ROLE_THEATRE_OWNER). Always present.
          example: "00000000-0000-4000-d000-000000000001"

    # ── Internal — Seat Lock/Release/Confirm ──────────────
...
    TheatreCreateRequest:
      type: object
      required:
        - name
        - cityId
        - address
        - ownerId
      properties:
        name:
          type: string
          minLength: 1
          example: PVR Cinemas
        cityId:
          type: integer
          minimum: 1
          example: 1
        address:
          type: string
          minLength: 1
          example: Phoenix MarketCity, Velachery
        latitude:
          type: number
          format: double
          example: 12.9716
        longitude:
          type: number
          format: double
          example: 77.5946
        ownerId:
          type: string
          format: uuid
          description: User ID of the theatre owner (ROLE_THEATRE_OWNER). Required — every theatre must have an owner.
          example: "00000000-0000-4000-d000-000000000001"
    TheatreUpdateRequest:
      type: object
      description: All fields optional; only provided fields are updated.
      properties:
        name:
          type: string
          minLength: 1
          example: PVR Cinemas
        cityId:
          type: integer
          minimum: 1
          example: 1
        address:
          type: string
          minLength: 1
          example: Phoenix MarketCity, Velachery
        latitude:
          type: number
          format: double
          example: 12.9716
        longitude:
          type: number
          format: double
          example: 77.5946
        ownerId:
          type: string
          format: uuid
          description: Reassign the theatre owner (admin-only). Must be a valid user UUID.
          example: "00000000-0000-4000-d000-000000000001"

    # ── Admin: Screens ──────────────────────────────────────
    Screen:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-4000-c000-000000000001"
        name:
          type: string
          example: Screen 3
        theatreId:
          type: string
          format: uuid
          example: "00000000-0000-4000-b000-000000000001"
        formatId:
          type: integer
          example: 2
        formatName:
          type: string
          example: 3D
        totalSeats:
          type: integer
          example: 150
    ScreenCreateRequest:
      type: object
      required:
        - name
        - formatId
        - totalSeats
      properties:
        name:
          type: string
          minLength: 1
          example: Screen 3
        formatId:
          type: integer
          minimum: 1
          example: 2
        totalSeats:
          type: integer
          minimum: 1
          example: 150
    ScreenUpdateRequest:
      type: object
      description: All fields optional; only provided fields are updated.
      properties:
        name:
          type: string
          minLength: 1
          example: Screen 3
        formatId:
          type: integer
          minimum: 1
          example: 2
        totalSeats:
          type: integer
          minimum: 1
          example: 150

    # ── Admin: Seats (inventory) ────────────────────────────
    TheatreSeat:
      type: object
      description: A physical seat on a screen (inventory-level, no booking status).
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        screenId:
          type: string
          format: uuid
          example: "00000000-0000-4000-c000-000000000001"
        rowLabel:
          type: string
          example: "A"
        seatNumber:
          type: integer
          example: 5
    SeatRowSpec:
      type: object
      description: Generates seats numbered 1..seatCount for the given row.
      required:
        - rowLabel
        - seatCount
      properties:
        rowLabel:
          type: string
          minLength: 1
          example: "A"
        seatCount:
          type: integer
          minimum: 1
          example: 15
    SeatSpec:
      type: object
      description: Explicit seat at (rowLabel, seatNumber).
      required:
        - rowLabel
        - seatNumber
      properties:
        rowLabel:
          type: string
          minLength: 1
          example: "A"
        seatNumber:
          type: integer
          minimum: 1
          example: 5
    SeatBulkCreateRequest:
      type: object
      description: >
        Provide either `rows` (sequential numbering per row) or `seats`
        (explicit pairs). Exactly one must be non-empty.
      properties:
        rows:
          type: array
          items:
            $ref: '#/components/schemas/SeatRowSpec'
          example:
            - rowLabel: "A"
              seatCount: 15
            - rowLabel: "B"
              seatCount: 15
        seats:
          type: array
          items:
            $ref: '#/components/schemas/SeatSpec'
          example:
            - rowLabel: "A"
              seatNumber: 1
            - rowLabel: "A"
              seatNumber: 2
    SeatListResponse:
      type: object
      properties:
        seats:
          type: array
          items:
            $ref: '#/components/schemas/TheatreSeat'

    # ── Admin: Shows ────────────────────────────────────────
...